+++ src/ir/global.cpp
@@ -17,7 +17,7 @@ void GlobalVariable::print(std::ostream& os) const {
   os << *(baseType()) << " ";
 
   if (isArray()) {
-    int idx = 0, dimensions = dims_cnt();
+    size_t idx = 0, dimensions = dims_cnt();
     print_ArrayInit(os, dimensions, 0, &idx);
   } else {
     os << *scalarValue();
@@ -27,17 +27,17 @@ void GlobalVariable::print(std::ostream& os) const {
 }
 
 void GlobalVariable::print_ArrayInit(std::ostream& os,
-                                     const int dimension,
-                                     const int begin,
-                                     int* idx) const {
+                                     const size_t dimension,
+                                     const size_t begin,
+                                     size_t* idx) const {
   auto atype = dyn_cast<ArrayType>(baseType());
   auto baseType = atype->baseType();
   bool is_zero = true;
   if (begin + 1 == dimension) {
-    int num = atype->dim(begin);
+    size_t num = atype->dim(begin);
 
     is_zero = true;
-    for (int i = 0; i < num; i++) {
+    for (size_t i = 0; i < num; i++) {
       auto cinit = dyn_cast<ir::Constant>(init(*idx + i));
       if (!cinit->is_zero()) {
         is_zero = false;
@@ -49,7 +49,7 @@ void GlobalVariable::print_ArrayInit(std::ostream& os,
       os << "zeroinitializer";
     else {
       os << "[";
-      for (int i = 0; i < num - 1; i++) {
+      for (size_t i = 0; i < num - 1; i++) {
         os << *(baseType) << " " << *init(*idx + i) << ", ";
       }
       os << *(baseType) << " " << *init(*idx + num - 1);
@@ -59,12 +59,12 @@ void GlobalVariable::print_ArrayInit(std::ostream& os,
   } else if (dimension == 2 + begin) {
     os << "[";
 
-    int num1 = atype->dim(begin), num2 = atype->dim(begin + 1);
-    for (int i = 0; i < num1 - 1; i++) {
+    size_t num1 = atype->dim(begin), num2 = atype->dim(begin + 1);
+    for (size_t i = 0; i < num1 - 1; i++) {
       os << "[" << num2 << " x " << *(baseType) << "] ";
 
       is_zero = true;
-      for (int j = 0; j < num2; j++) {
+      for (size_t j = 0; j < num2; j++) {
         auto cinit = dyn_cast<ir::Constant>(init(*idx + i * num2 + j));
         if (!cinit->is_zero()) {
           is_zero = false;
@@ -75,7 +75,7 @@ void GlobalVariable::print_ArrayInit(std::ostream& os,
         os << "zeroinitializer, ";
       else {
         os << "[";
-        for (int j = 0; j < num2 - 1; j++) {
+        for (size_t j = 0; j < num2 - 1; j++) {
           os << *(baseType) << " " << *init(*idx + i * num2 + j) << ", ";
         }
         os << *(baseType) << " " << *init(*idx + i * num2 + num2 - 1);
@@ -85,7 +85,7 @@ void GlobalVariable::print_ArrayInit(std::ostream& os,
     os << "[" << num2 << " x " << *(baseType) << "] ";
 
     is_zero = true;
-    for (int j = 0; j < num2; j++) {
+    for (size_t j = 0; j < num2; j++) {
       auto cinit = dyn_cast<ir::Constant>(init(*idx + (num1 - 1) * num2 + j));
       if (!cinit->is_zero()) {
         is_zero = false;
@@ -96,7 +96,7 @@ void GlobalVariable::print_ArrayInit(std::ostream& os,
       os << "zeroinitializer";
     else {
       os << "[";
-      for (int j = 0; j < num2 - 1; j++) {
+      for (size_t j = 0; j < num2 - 1; j++) {
         os << *(baseType) << " " << *init(*idx + (num1 - 1) * num2 + j) << ", ";
       }
       os << *(baseType) << " " << *init(*idx + (num1 - 1) * num2 + num2 - 1);
@@ -108,26 +108,26 @@ void GlobalVariable::print_ArrayInit(std::ostream& os,
   } else {
     os << "[";
 
-    int num = atype->dim(begin), num1 = atype->dim(begin + 1),
+    size_t num = atype->dim(begin), num1 = atype->dim(begin + 1),
         num2 = atype->dim(begin + 2);
-    for (int i = 1; i < num; i++) {
-      for (int cur = begin + 1; cur < dimension; cur++) {
+    for (size_t i = 1; i < num; i++) {
+      for (size_t cur = begin + 1; cur < dimension; cur++) {
         auto value = atype->dim(cur);
         os << "[" << value << " x ";
       }
       os << *(baseType);
-      for (int cur = begin + 1; cur < dimension; cur++)
+      for (size_t cur = begin + 1; cur < dimension; cur++)
         os << "]";
       os << " ";
       print_ArrayInit(os, dimension, begin + 1, idx);
       os << ", ";
     }
-    for (int cur = begin + 1; cur < dimension; cur++) {
+    for (size_t cur = begin + 1; cur < dimension; cur++) {
       auto value = atype->dim(cur);
       os << "[" << value << " x ";
     }
     os << *(baseType);
-    for (int cur = begin + 1; cur < dimension; cur++)
+    for (size_t cur = begin + 1; cur < dimension; cur++)
       os << "]";
     os << " ";
     print_ArrayInit(os, dimension, begin + 1, idx);
